Cleaning up the Developing section;
A lot of these changes have more to do with the move to Docusaurus.
Links to descendant changes are now using the DocCardList component
rather than linking to each individual page under the main page. Some of
these links were also removed as they added unnecessary steps to reading
the documentation which is more streamlined than reading on GitHub.